Refactor remaining items from psng.py
Split off each of these into their own files:

* Length Measurements
* Tool Measurements
* Workpiece Measurements
