+++ psng/python/__init__.py
@@ -18,9 +18,12 @@
 import linuxcnc  # to get our own error system
 
 from .jog import ProbeScreenJog
+from .length_measurement import ProbeScreenLengthMeasurement
 from .psng import ProbeScreenClass
 from .rotation import ProbeScreenRotation
 from .settings import ProbeScreenSettings
+from .tool_measurement import ProbeScreenToolMeasurement
+from .workpiece_measurement import ProbeScreenWorkpieceMeasurement
 from .zero import ProbeScreenZero
 
 
@@ -31,4 +34,7 @@ def get_handlers(halcomp, builder, useropts):
         ProbeScreenJog(halcomp, builder, useropts),
         ProbeScreenZero(halcomp, builder, useropts),
         ProbeScreenRotation(halcomp, builder, useropts),
+        ProbeScreenToolMeasurement(halcomp, builder, useropts),
+        ProbeScreenWorkpieceMeasurement(halcomp, builder, useropts),
+        ProbeScreenLengthMeasurement(halcomp, builder, useropts),
     ]
